requirements.txt
blinker==1.9.0
certifi==2026.2.25
cffi==2.0.0
click==8.3.1
cloudinary==1.44.1
colorama==0.4.6
cryptography==46.0.5
Flask==3.1.3
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
pycparser==3.0
PyMySQL==1.1.2
python-dotenv==1.2.2
six==1.17.0
urllib3==2.6.3
Werkzeug==3.1.6
